Clean up configs
