+++ @zapp-framework/core/src/working_tree/props/TextConfig.ts
@@ -1,11 +1,11 @@
-import { ConfigBuilder } from './Config.js'
+import { BaseConfigBuilder } from './BaseConfig.js'
 import { Alignment } from './types.js'
 
 export function TextConfig(id: string) {
   return new TextConfigBuilder(id)
 }
 
-export class TextConfigBuilder extends ConfigBuilder {
+export class TextConfigBuilder extends BaseConfigBuilder {
   public textColor(textColor: number) {
     this.config.textColor = textColor
     return this
