+++ @zapp-framework/core/src/working_tree/views/Arc.ts
@@ -1,10 +1,9 @@
-import type { RequireSome } from '../../utils.js'
 import { ViewNode } from '../ViewNode.js'
 import { WorkingTree } from '../WorkingTree.js'
 import { NodeType } from '../../NodeType.js'
 import { ConfigBuilder } from '../props/Config.js'
 
-export function Arc(configBuilder: RequireSome<ConfigBuilder, 'build'>) {
+export function Arc(configBuilder: ConfigBuilder) {
   const config = configBuilder.build()
   const current = WorkingTree.current as ViewNode
