+++ @zapp-framework/core/src/working_tree/views/BareText.ts
@@ -1,10 +1,9 @@
-import type { RequireSome } from '../../utils.js'
 import { ViewNode } from '../ViewNode.js'
 import { WorkingTree } from '../WorkingTree.js'
 import { NodeType } from '../../NodeType.js'
 import { ConfigBuilder } from '../props/Config.js'
 
-export function BareText(configBuilder: RequireSome<ConfigBuilder, 'build'>, text: string) {
+export function BareText(configBuilder: ConfigBuilder, text: string) {
   const config = configBuilder.build()
   config.text = text
   const current = WorkingTree.current as ViewNode
