+++ @zapp-framework/core/src/working_tree/views/Image.ts
@@ -1,10 +1,9 @@
-import type { RequireSome } from '../../utils.js'
 import { ViewNode } from '../ViewNode.js'
 import { WorkingTree } from '../WorkingTree.js'
 import { NodeType } from '../../NodeType.js'
 import { ConfigBuilder } from '../props/Config.js'
 
-export function Image(configBuilder: RequireSome<ConfigBuilder, 'build'>, source: string) {
+export function Image(configBuilder: ConfigBuilder, source: string) {
   const config = configBuilder.build()
   config.source = source
   const current = WorkingTree.current as ViewNode
