+++ @zapp-framework/core/src/working_tree/views/Row.ts
@@ -1,10 +1,9 @@
-import type { RequireSome } from '../../utils.js'
 import { ViewNode } from '../ViewNode.js'
 import { WorkingTree } from '../WorkingTree.js'
 import { NodeType } from '../../NodeType.js'
 import { ConfigBuilder } from '../props/Config.js'
 
-export function Row(configBuilder: RequireSome<ConfigBuilder, 'build'>, body?: () => void) {
+export function Row(configBuilder: ConfigBuilder, body?: () => void) {
   const config = configBuilder.build()
   const current = WorkingTree.current as ViewNode
