+++ @zapp-framework/core/src/working_tree/views/Screen.ts
@@ -1,12 +1,11 @@
-import type { RequireSome } from '../../utils.js'
 import { ViewNode } from '../ViewNode.js'
 import { WorkingTree } from '../WorkingTree.js'
 import { NodeType } from '../../NodeType.js'
-import { ConfigBuilder } from '../props/Config.js'
 import { Zapp } from '../../ZappInterface.js'
+import { ConfigBuilder } from '../props/Config.js'
 
 export function ScreenBody(
-  configBuilder: RequireSome<ConfigBuilder, 'build'>,
+  configBuilder: ConfigBuilder,
   body?: (params?: Record<string, unknown>) => void
 ) {
   const config = configBuilder.build()
@@ -30,23 +29,20 @@ export function ScreenBody(
 }
 
 let simpleScreenImplementation = (
-  configBuilder: RequireSome<ConfigBuilder, 'build'>,
+  configBuilder: ConfigBuilder,
   body?: (params?: Record<string, unknown>) => void
 ) => {
   ScreenBody(configBuilder, body)
 }
 
 export function setSimpleScreenImplementation(
-  implementation: (
-    config: RequireSome<ConfigBuilder, 'build'>,
-    body?: (params?: Record<string, unknown>) => void
-  ) => void
+  implementation: (config: ConfigBuilder, body?: (params?: Record<string, unknown>) => void) => void
 ) {
   simpleScreenImplementation = implementation
 }
 
 export function SimpleScreen(
-  configBuilder: RequireSome<ConfigBuilder, 'build'>,
+  configBuilder: ConfigBuilder,
   body?: (params?: Record<string, unknown>) => void
 ) {
   simpleScreenImplementation(configBuilder, body)
